cheri: Fix errors in drivers/usb/
- uintptr_t vs. unsigned long and ioctl issues.
- Unpack a struct in drivers/usb/gadget/function/uvc_configfs.h
  that contains pointers for CHERI.

+++ drivers/usb/phy/phy-mv-usb.c
@@ -748,7 +748,7 @@ static int mv_otg_probe(struct platform_device *pdev)
 	}
 
 	mvotg->op_regs =
-		(struct mv_otg_regs __iomem *) ((unsigned long) mvotg->cap_regs
+		(struct mv_otg_regs __iomem *) ((uintptr_t) mvotg->cap_regs
 			+ (readl(mvotg->cap_regs) & CAPLENGTH_MASK));
 
 	if (pdata->id) {
